refactor(signals): focus store, drop the paywall, split the readout
Strip focus moves into a small zustand store. Every node used to take
`selectedId` and `hoveredId` as props, so nothing could ever bail out of
a re-render and one hover re-rendered all 41 nodes. Nodes now subscribe
to their own three booleans with a shallow selector.

Selection is NOT in the store: `useClusterId` stays the only source of
truth and is passed to the selectors, matching `getChartClusters`.

`memo` alone bought nothing — leaving a band writes `null` before
entering the next one writes its id, and a null focus un-mutes every
band. Hover is delegated on the strip instead, one `pointerover` reading
`data-cluster-band`, so a band-to-band move is one write and re-renders
two nodes instead of forty-one.

With hover out of the section, the card no longer re-renders on it, so
the memoised-element workaround and its ten-entry dep list are gone.

The clusters paywall is removed. Note the API still rewrites cluster
names for Free/Hobby workspaces; that is untouched here.

`cluster-readout.tsx` and the icicle node are split into folders, with
the hover-card lifecycle isolated in `use-hover-card`.
